requirements.txt
crewai
streamlit
python-dotenv
requests
PyPDF2
matplotlib